requirements.txt
# Core dependencies
pydantic>=2.5.0
pydantic-settings>=2.1.0
fastapi>=0.104.0
uvicorn>=0.24.0
python-dotenv>=1.0.0

# ML and NLP
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
spacy>=3.7.0
scikit-learn>=1.3.0
numpy>=1.24.0

# LLM clients
google-generativeai>=0.3.0

# Database
sqlalchemy>=2.0.0
alembic>=1.12.0

# CLI and utilities
click>=8.1.0
rich>=13.7.0
python-dateutil>=2.8.2

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0
pytest-mock>=3.12.0
httpx>=0.25.0

# Development
black>=23.11.0
flake8>=6.1.0
mypy>=1.7.0
isort>=5.12.0
